+++ pyccel/codegen/make_pipeline.py
@@ -106,14 +106,6 @@ def execute_pyccel_make(files, *,
     base_dirpath = os.getcwd()
     sys.path.insert(0, base_dirpath)
 
-    def handle_error(stage):
-        """
-        Unified way to handle errors: print formatted error message.
-        Caller should then raise exception.
-        """
-        print(f'\nERROR at {stage} stage')
-        errors.check()
-
     # Define working directory 'folder'
     if folder is None or folder == "":
         folder = Path(os.getcwd())
@@ -146,6 +138,17 @@ def execute_pyccel_make(files, *,
     start_syntax = time.time()
     timers["Initialisation"] = start_syntax-start
 
+    if len(files) == 0:
+        errors.report("No files passed to pyccel make",
+                      severity='fatal')
+    else:
+        print("Translating :")
+        for f in files:
+            print("-", f)
+
+    cwd = os.getcwd()
+    files = [f.relative_to(cwd) if f.is_absolute() else f for f in files]
+
     parsers = {f: Parser(f.absolute(), output_folder = folder) for f in files}
 
     to_remove = []
@@ -155,17 +158,8 @@ def execute_pyccel_make(files, *,
             to_remove.append(f)
             continue
         # Parse Python file
-        try:
-            p.parse(verbose=verbose, d_parsers_by_filename = {f.absolute(): p for f, p in parsers.items()})
-        except NotImplementedError as error:
-            errors.report(str(error)+'\n'+PYCCEL_RESTRICTION_TODO,
-                severity='error',
-                traceback=error.__traceback__)
-        except PyccelError:
-            handle_error('parsing (syntax)')
-            raise
+        p.parse(verbose=verbose, d_parsers_by_filename = {f.absolute(): p for f, p in parsers.items()})
         if errors.has_errors():
-            handle_error('parsing (syntax)')
             raise PyccelSyntaxError('Syntax step failed')
 
     for r in to_remove:
@@ -176,19 +170,9 @@ def execute_pyccel_make(files, *,
     start_semantic = time.time()
     # Annotate abstract syntax Tree
     for f, p in parsers.items():
-        try:
-            p.annotate(verbose = verbose)
-        except NotImplementedError as error:
-            errors.report(str(error)+'\n'+PYCCEL_RESTRICTION_TODO,
-                severity='error',
-                traceback=error.__traceback__)
-        except PyccelError:
-            handle_error('annotation (semantic)')
-            # Raise a new error to avoid a large traceback
-            raise PyccelSemanticError('Semantic step failed') from None
+        p.annotate(verbose = verbose)
 
         if errors.has_errors():
-            handle_error('annotation (semantic)')
             raise PyccelSemanticError('Semantic step failed')
 
     timers["Semantic Stage"] = time.time() - start_semantic
@@ -213,19 +197,9 @@ def execute_pyccel_make(files, *,
         fname = (pyccel_dirpath / f).with_suffix('')
         output_dir = fname.parent
         os.makedirs(output_dir, exist_ok=True)
-        try:
-            fname, prog_name = codegen.export(fname)
-        except NotImplementedError as error:
-            errors.report(str(error)+'\n'+PYCCEL_RESTRICTION_TODO,
-                severity='error',
-                traceback=error.__traceback__)
-        except PyccelError:
-            handle_error('code generation')
-            # Raise a new error to avoid a large traceback
-            raise PyccelCodegenError('Code generation failed') from None
+        fname, prog_name = codegen.export(fname)
 
         if errors.has_errors():
-            handle_error('code generation')
             raise PyccelCodegenError('Code generation failed')
 
         timers["Codegen Stage"] += time.time() - start_codegen
@@ -240,22 +214,13 @@ def execute_pyccel_make(files, *,
         else:
             start_wrapper_creation = time.time()
             wrappergen = Wrappergen(codegen, codegen.name, language, verbose)
-            try:
-                wrappergen.wrap(str((base_dirpath / f).parent))
-            except PyccelError:
-                handle_error('code generation (wrapping)')
-                raise
+            wrappergen.wrap(str((base_dirpath / f).parent))
             timers['Wrapper creation'] += time.time() - start_wrapper_creation
 
             start_wrapper_printing = time.time()
-            try:
-                wrapper_files = wrappergen.print(output_dir)
-            except PyccelError:
-                handle_error('code generation (wrapping)')
-                raise
+            wrapper_files = wrappergen.print(output_dir)
 
             if errors.has_errors():
-                handle_error('code generation (wrapping)')
                 raise PyccelCodegenError('Code generation step failed')
 
             timers['Wrapper printing'] += time.time() - start_wrapper_printing
@@ -292,38 +257,21 @@ def execute_pyccel_make(files, *,
                             language = language, verbose = verbose, convert_only = True, mod_obj = t,
                             installed_libs = stdlib_deps)
 
-    try:
-        build_project = BuildProject(base_dirpath, targets.values(), printed_languages,
-                                     stdlib_deps)
+    build_project = BuildProject(base_dirpath, targets.values(), printed_languages,
+                                 stdlib_deps)
 
-        build_sys = build_system_handler[build_system](pyccel_dirpath, base_dirpath, folder,
-                                                       verbose = verbose, debug_mode = debug,
-                                                       compiler = compiler, accelerators = accelerators)
+    build_sys = build_system_handler[build_system](pyccel_dirpath, base_dirpath, folder,
+                                                   verbose = verbose, debug_mode = debug,
+                                                   compiler = compiler, accelerators = accelerators)
 
-        build_sys.generate(build_project)
-    except NotImplementedError as error:
-        errors.report(str(error)+'\n'+PYCCEL_RESTRICTION_TODO,
-            severity='error',
-            traceback=error.__traceback__)
-    except PyccelError:
-        handle_error('code generation')
-        # Raise a new error to avoid a large traceback
-        raise PyccelCodegenError('Code generation failed') from None
+    build_sys.generate(build_project)
 
     timers['Build system printing'] = time.time() - start_build_system_printing
-
-    if errors.has_errors():
-        handle_error('build system generation')
-        raise PyccelCodegenError('Build system generation failed')
 
     if build_code:
         start_compilation = time.time()
         build_sys.compile()
         timers['Compilation'] = time.time() - start_compilation
-
-    # Print all warnings now
-    if errors.has_warnings():
-        errors.check()
 
     # Change working directory back to starting point
     pyccel_stage.pyccel_finished()