+++ pyccel/commands/epyccel.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 #------------------------------------------------------------------------------------------#
 # This file is part of Pyccel which is released under MIT License. See the LICENSE file or #
 # go to https://github.com/pyccel/pyccel/blob/devel/LICENSE for full license details.      #
@@ -17,6 +16,7 @@ from importlib.machinery import ExtensionFileLoader
 
 from filelock import FileLock, Timeout
 
+from pyccel.utilities.stage    import PyccelStage
 from pyccel.utilities.strings  import random_string
 from pyccel.codegen.pipeline   import execute_pyccel
 from pyccel.errors.errors      import ErrorsMode, PyccelError, Errors
@@ -338,23 +338,30 @@ def epyccel_seq(function_class_or_module, *,
         with open(pymod_filename, 'w', encoding='utf-8') as f:
             f.writelines(code)
 
-        # Generate shared library
-        execute_pyccel(pymod_filename,
-                       verbose         = verbose,
-                       time_execution  = time_execution,
-                       language        = language,
-                       compiler_family = compiler_family_or_config,
-                       flags           = flags,
-                       wrapper_flags   = wrapper_flags,
-                       include         = include,
-                       libdir          = libdir,
-                       modules         = (),
-                       libs            = libs,
-                       debug           = debug,
-                       accelerators    = accelerators,
-                       output_name     = module_name,
-                       conda_warnings  = conda_warnings,
-                       context_dict    = context_dict)
+        pyccel_stage = PyccelStage()
+        try:
+            # Generate shared library
+            execute_pyccel(pymod_filename,
+                           verbose         = verbose,
+                           time_execution  = time_execution,
+                           language        = language,
+                           compiler_family = compiler_family_or_config,
+                           flags           = flags,
+                           wrapper_flags   = wrapper_flags,
+                           include         = include,
+                           libdir          = libdir,
+                           modules         = (),
+                           libs            = libs,
+                           debug           = debug,
+                           accelerators    = accelerators,
+                           output_name     = module_name,
+                           conda_warnings  = conda_warnings,
+                           context_dict    = context_dict)
+        except PyccelError as err:
+            raise err
+        finally:
+            pyccel_stage.pyccel_finished()
+            print(errors, end='')
 
 
         # Import shared library